refactor: extract UserContext from GlobalLoading to wrapRootElement
This allows the UserContext provider to sit at the highest level in the tree,
making values available to every page that needs it without requiring the page
to contort itself to do work after the <Layout> has rendered. This is mostly
useful to pages created in src/pages which want to make use of user/demo API
keys.
