+++ gatsby-browser.ts
@@ -40,4 +40,10 @@ const shouldUpdateScroll = ({ prevRouterProps, routerProps: { location } }) => {
   );
 };
 
-export { onClientEntry, shouldUpdateScroll };
+/**
+ * Load our user state
+ */
+import UserContextWrapper from 'src/contexts/user-context/wrap-with-provider';
+const wrapRootElement = UserContextWrapper;
+
+export { onClientEntry, shouldUpdateScroll, wrapRootElement };
