+++ gatsby-ssr.ts
@@ -1,0 +1,7 @@
+/**
+ * Load our user state
+ */
+import UserContextWrapper from 'src/contexts/user-context/wrap-with-provider';
+const wrapRootElement = UserContextWrapper;
+
+export { wrapRootElement };
